// src/dungeonslayers/model/ds_class.ts
export enum DSClass {
    KRI = 'KRI',
    SPÄ = 'SPÄ',
    ZAW = 'ZAW',
    Hei = 'Hei',
    Sch = 'Sch',
    Zau = 'Zau',
    BER = 'BER',
    PAL = 'PAL',
    WAM = 'WAM',
    ATT = 'ATT',
    MDB = 'MDB',
    WDL = 'WDL',
    DRU = 'DRU',
    KMÖ = 'KMÖ',
    KLE = 'KLE',
    ELE = 'ELE',
    ERZ = 'ERZ',
    KRZ = 'KRZ',
    BLU = 'BLU',
    DÄM = 'DÄM',
    NEK = 'NEK',
}